+++ src/clj_commons/pretty_impl.clj
@@ -12,3 +12,8 @@
 (def ^:const csi
   "The control sequence initiator: `ESC [`"
   "\u001b[")
+
+(def align->pad
+  {:left :right
+   :right :left
+   :center :both})